kserve/message_sidecar/kafka_sink/requirements.txt
kafka-python
kserve
pillow>=10.3.0,<11.0.0
cloudevents
fastapi